A hack to ease compatibility with pre-2.3 Pythons:  by default, doctest now accepts "True" when a test expects "1", and similarly for "False" versus "0".  This is un-doctest-like, but on balance makes it much more pleasant to write doctests that pass under 2.2 and 2.3.  I expect it to go away again, when 2.2 is forgotten.  In the meantime, there's a new doctest module constant that can be passed to a new optional argument, if you want to turn this behavior off.
Note that this substitution is very simple-minded:  the expected and
actual outputs have to consist of single tokens.  No attempt is made,
e.g., to accept [True, False] when a test expects [1, 0].  This is a
simple hack for simple tests, and I intend to keep it that way.

--HG--
branch : legacy-trunk