FreeBSD 5.x has moved some library routines and typedefs outside the scope of the _XOPEN_SOURCE and _POSIX_C_SOURCE symbols, including: - getloadavg() - typedefs for u_int, u_long, u_char, u_short, ushort & uint
These are now all defined under the control of a __BSD_VISIBLE symbol.

The lack of the typedefs causes several extension modules to build
incorrectly or not at all, and is the cause of failures reported for
test_socket and test_tempfile on this platform
(see python-dev: 29/6/03, pieterb@gewis.nl, "Running tests on freebsd5")

This change does not appear to be needed in the 2.2 branch.

--HG--
branch : legacy-trunk