Fixes for SF #1076485, which I'll apply to the CVS head too.  The problem was caused by a self._input.readline() call that wasn't checking for the NeedsMoreData marker.
msg_43.txt contains a message that illustrates the problem, when
email.message_from_*() is called.  That interface uses the Parser API, which
splits reads into 8192 byte chunks.  It so happens that for the test message,
the 8192 chunk falls inside a message/delivery-status, which is where in the
FeedParser the readline() call was that didn't check for NeedsMoreData.

I also added an assert to unreadline() so it'll be more evident if an attempt
to push back NeedsMoreData ever happens again.

Bump the email package version number.

--HG--
branch : legacy-trunk